arm-common: Avoid map/unmap of empty_page in iommu
Jailhouse maps entire hypervisor memory such that
when accessed from VMs address space, it hits an empty page.
This is used as a guard to avoid faults while
shutting down jailhouse.

We don't need to map/unmap the hypervisor memory to
"empty_page" in the iommu address space as well.
This wastes lot of iommu mapping unnecessarily.
Skip the iommu map/unmap for the empty_page.

Without any mapping, there will be a fault, which is okay.
No VM is supposed to setup DMA in the hypervisor reserved memory.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>