configs: arm64: k3: Describe streamIDs for K3 cells
Add streamIDs in the root cell and inmate cell.
This will make sure that the memory regions are mapped
correctly in the IO space using an IOMMU driver.

Even before the cells are created, all devices are used by root cell
Make sure that streamIDs for those devices are also configured with
valid translation tables.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>