configs: j721e-evm: Do not map inmate memory as MEM_DMA
Do not mark the memory reserved for inmate as DMA capable.
Split the second bank into two.
First 0.5G is for root cell, mark it as MEM_DMA
Next 1.5G is for inmate cell, mark it only MEM_LOADABLE, avoid MEM_DMA

This fixes the broken I/O isolation for PVU IOMMU driver.
PVU will not map the inmate memory in the root cell. This works
around the limitation of not able to unmap the inmate memory.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>