core: Make streamIDs a resizeable array
This change was suggested in the upstream review of the SMMUv3 driver.
Also, rename it to stream_ids for uniformity.

Stream ID array no longer needs to be terminated with
JAILHOUSE_INVALID_STREAMID, because we have a size for it already.

While here, fix num_regmaps being in the wrong place in
jailhouse-cell-linux. num_regmaps comes after pci_caps in the cell
config, but here, it is after num_irqchips instead.

Signed-off-by: Pratyush Yadav <p-yadav1@ti.com>