arm64: iommu: smmu-v3: Fix queue_error() incorrectly checking for error
Right now queue_error() checks the lowest bit of the consumer to see if
there is an error. This is not correct. The correct indicators of a queue
error are the gerror and gerrorn registers.

Add gerr_mask to struct arm_smmu_queue which indicates which bits being
toggled in gerror indicate an error condition for that queue.

While here, fix a double semi-colon in arm_smmu_init_one_queue().

Signed-off-by: Pratyush Yadav <p-yadav1@ti.com>