configs: arm64: k3-j721e-evm: Fix failure with PCIe
PCI kernel framework requires that all the instances in the
device tree either specify the PCIe domain or none does.

Currently, Jailhouse dynamic overlay describes the PCI domain
which causes problems because root cell DTS does not describe
this.

Fix this by not specifying the domain in root cell config.
Also, fix the size of the PCI target region to match with the
correct description in TRM.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>