configs: k3-j721e-evm: Describe separate IVSHMEM regions for demo
Current IVSHMEM regions are described to be used for ivshmem-net
driver. For standalone ivshmem communication, these regions need to
be added explicitly instead of using the macro.

Add regions for a 2 peer IVSHMEM communication
 (0 = root cell, 1 = baremetal / linux-demo)

Also change the protocol to UNDEFINED so that the ivshmem-net driver
does not get engaged. Switch bdf = 0 for consistency across all
platforms.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>