configs: arm64: k3-am625-sk: Switch inmate boot console alias to serial3
Traditionally the serial3 DTS alias is used for the 2nd main-domain UART
on K3 devices so go ahead and update the alias and its usage accordingly
for better consistency.

While at it also do a minor cleanup of the UART's DTS node to use the
TISCI standard definitions as used in the corresponding Kernel DTS files.

Upstream-Status: Submitted [jailhouse-dev@googlegroups.com]

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>