Add crash test application for linux instance
This crash test application can be used to send icmp packets to test
whether either cell is working (root cell or inmate cell). It
relies on ivshmem net module which makes a virtual interface on both
root and inmate side.

After assigning an ipaddress to the virtual interface, root or inmate can
send packets to each other. This application provides a sleep functionality
and some other arguments which can be used to control the number and interval
of sending the packets.

Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>