configs: arm64: k3-{am6x/j7x}: Set pci_domain config to -1
Hard-coded PCI domains are not released on device removal. If the PCI device is removed and added a second time in the same boot, it will lead to failure. This is because the same unreleased hard-coded domain is requested the second time on device addition.

Hence set the pci_domain config to -1 (PCI_DOMAIN_NR_NOT_SET) in order to ensure that the domain is allocated dynamically and is automatically released on device removal, thereby allowing multiple device additions and removals in a single boot.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>